moved environmentNotice to shared component and made some refactors
